+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## [0.2.1] - 2019-09-13
+### Changed
+- Fixed example in README
+- Fixed bug with named groups in URL only processing first group
+
 ## [0.2.0] - 2019-02-21
 ### Added
 - New function `create-request-method-handler`
@@ -13,5 +18,6 @@ All notable changes to this project will be documented in this file. This change
 ### Added
 - Initial public release
 
+[0.2.1]: https://github.com/sbreatnach/treo/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/sbreatnach/treo/compare/0.1.0...0.2.0
 [Unreleased]: https://github.com/sbreatnach/treo/compare/0.2.0...HEAD